# Logs Panel: Performance improvement in logs rows