**What is this feature?**
Flamegraph of each re-render in dashboard showed that a majority of each re-render was spent in the LogsRows, and re-rendering was quite expensive. Currently the LogRows re-renders whenever any of the parent components change.

<img width="1728" alt="image" src="https://github.com/grafana/grafana/assets/109082771/6b400f2b-7742-4d79-b58e-51c136905f10">


With this fix in place we no longer re-render the logRows component unless the log rows change. Cutting out a significant number of unnecessary re-renders.
<img width="1727" alt="image" src="https://github.com/grafana/grafana/assets/109082771/a24737b3-8ce2-486a-8d3c-c10673c589cc">

Using profiles gathered when resizing a logs panel:
Most renders without this fix take 15ms-20ms, max 19.6ms, min 14.5
On this branch I'm seeing most renders at <1ms, max is 2.3ms, min 0.6ms

Potentially fixes https://github.com/grafana/grafana/issues/53016

**Why do we need this feature?**
Logs Panel can be slow, especially in a dashboard context, this is especially apparent in dashboards with many logs panels, or when resizing a logs panel in a dashboard.

**Who is this feature for?**
All logs users.

**Special notes for your reviewer:**
I would expect more bugs, but it looks like all of the other props that we'd expect we want to re-render on (showDuplicates, logSortOrder (although this appears unused by the LogRow component?), themselves change the log rows, which triggers a single re-render now, instead of several. 

Please check that:
- [ ] It works as expected from a user's perspective.
- [ ] If this is a pre-GA feature, it is behind a feature toggle.
- [ ] The docs are updated, and if this is a [notable improvement](https://grafana.com/docs/writers-toolkit/writing-guide/contribute-release-notes/#how-to-determine-if-content-belongs-in-a-whats-new-document), it's added to our [What's New](https://grafana.com/docs/writers-toolkit/writing-guide/contribute-release-notes/) doc.